only acquire mergelock once
